requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
-e file:.
    # via -r requirements.in
astor==0.8.1
    # via flake8-qgis
attrs==25.1.0
    # via pytest-subtests
certifi==2025.1.31
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
coverage[toml]==7.6.12
    # via pytest-cov
distlib==0.3.9
    # via virtualenv
filelock==3.17.0
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements.in
    #   flake8-pyproject
    #   flake8-qgis
    #   flake8-spellcheck
flake8-pyproject==1.2.3
    # via -r requirements.in
flake8-qgis==1.0.0
    # via -r requirements.in
flake8-spellcheck==0.28.0
    # via -r requirements.in
identify==2.6.8
    # via pre-commit
idna==3.10
    # via requests
importlib-metadata==8.6.1
    # via qgis-plugin-dev-tools
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.13.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via
    #   pytest
    #   qgis-plugin-dev-tools
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via
    #   pytest
    #   pytest-qt
pre-commit==3.4.0
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyqt5-stubs==5.15.6.0
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-dotenv
    #   pytest-mock
    #   pytest-qgis
    #   pytest-qt
    #   pytest-subtests
    #   pytest-timeout
pytest-cov==6.0.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
pytest-qgis==2.1.0
    # via -r requirements.in
pytest-qt==4.4.0
    # via -r requirements.in
pytest-subtests==0.14.1
    # via -r requirements.in
pytest-timeout==2.3.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   pytest-dotenv
    #   qgis-plugin-dev-tools
pyyaml==6.0.2
    # via pre-commit
qgis-plugin-dev-tools==0.9.1
    # via -r requirements.in
qgis-plugin-tools==0.5.0
    # via profiler-qgis-plugin
requests==2.32.3
    # via qgis-plugin-dev-tools
ruff==0.9.10
    # via -r requirements.in
tomli==2.2.1
    # via qgis-plugin-dev-tools
typing-extensions==4.12.2
    # via mypy
urllib3==2.3.0
    # via requests
virtualenv==20.29.3
    # via pre-commit
zipp==3.21.0
    # via importlib-metadata
